+++ packages/better_player/lib/src/controls/better_player_clickable_widget.dart
@@ -20,16 +20,13 @@ class BetterPlayerMaterialClickableWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Material(
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(60),
-      ),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(60)),
       clipBehavior: Clip.hardEdge,
       color: Colors.transparent,
       child: InkWell(
         onTap: () {
           PlayerLogger.debug(
             'Tapped on: $semanticsIdentifier ($semanticsLabel)',
-            tag: 'ClickableWidget',
           );
           onTap();
         },