+++ packages/better_player/lib/src/controls/better_player_cupertino_bottom_bar.dart
@@ -230,11 +230,7 @@ class _BetterPlayerCupertinoSkipButton extends StatelessWidget {
           color: Colors.transparent,
           margin: EdgeInsets.only(left: isBack ? 10 : 0, right: isBack ? 0 : 8),
           padding: EdgeInsets.symmetric(horizontal: isBack ? 8 : 6),
-          child: Icon(
-            icon,
-            color: iconColor,
-            size: barHeight * 0.4,
-          ),
+          child: Icon(icon, color: iconColor, size: barHeight * 0.4),
         ),
       ),
     );